requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers
django-allauth
factory-boy==3.2.1
Faker==13.13.0
mysqlclient==2.1.0
Pillow==9.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tblib==1.7.0
tzdata==2022.1
numpy
